Moved tests to neighbour-service
