+++ message-collector/src/test/resources/application.properties
@@ -1,12 +1,7 @@
-spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
-spring.datasource.url=jdbc:postgresql://localhost:7070/federation
-spring.datasource.username=federation
-spring.datasource.password=federation
-spring.jpa.hibernate.ddl-auto=create-drop
 
 # logging
 logging.pattern.console=%d{yyyy-MM-dd HH:mm:ss} [%-13.13thread] %-5level %logger{36} - %msg%n
-logging.level.org.hibernate.SQL=debug
+
 
 collector.localIxnDomainName=local
 collector.localIxnFederationPort =5671
