+++ neighbour-discoverer/src/test/java/no/vegvesen/ixn/federation/discoverer/NeighbourDiscovererIT.java
@@ -30,6 +30,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.*;
 
+//TODO: move relevant tests to NeighbourServiceIT
 @RunWith(SpringRunner.class)
 @SpringBootTest
 @ContextConfiguration(initializers = {PostgresTestcontainerInitializer.Initializer.class})
