+++ neighbour-service/pom.xml
@@ -28,21 +28,20 @@
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-test</artifactId>
             <scope>test</scope>
-            <exclusions>
-                <exclusion>
-                    <groupId>org.junit.vintage</groupId>
-                    <artifactId>junit-vintage-engine</artifactId>
-                </exclusion>
-            </exclusions>
         </dependency>
+
+        <dependency>
+            <groupId>org.testcontainers</groupId>
+            <artifactId>postgresql</artifactId>
+            <scope>test</scope>
+        </dependency>
+
         <dependency>
             <groupId>net.nordicway.interchange</groupId>
             <artifactId>neighbour-rest-facade</artifactId>
             <version>${project.version}</version>
             <scope>compile</scope>
         </dependency>
-
-
     </dependencies>
 
 </project>
