+++ neighbour-service/src/test/java/no/vegvesen/ixn/federation/NeighbourServiceITApplication.java
@@ -1,0 +1,12 @@
+package no.vegvesen.ixn.federation;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+/**
+ * Make sure to put the SpringBootApplication high enough in the package tree
+ * in order to auto-wire beans below its package.
+ */
+public class NeighbourServiceITApplication {
+
+}
