+++ neighbour-service/src/test/resources/application.properties
@@ -1,0 +1,8 @@
+spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.PostgreSQLDialect
+spring.datasource.url=jdbc:postgresql://localhost:7070/federation
+spring.datasource.username=federation
+spring.datasource.password=federation
+spring.jpa.hibernate.ddl-auto=create-drop
+
+
+interchange.node-provider.name=myName
